# static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "GPG Webservice API",
    "version": "1.0.0",
    "description": "A Flask-based GPG webservice that provides cryptographic operations (signing, verification, encryption, decryption) via HTTP endpoints. The service handles user registration, authentication, and automatic GPG key management with secure storage.",
    "contact": {
      "name": "GPG Webservice Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5555",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "API key for authentication (obtained during user registration)"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        }
      },
...
    }
  },
  "paths": {
    "/register": {
      "post": {
        "tags": ["User Management"],
        "summary": "Register a new user",
        "description": "Register a new user with automatic GPG keypair generation. Returns a one-time API key.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50,
                    "description": "Username (alphanumeric, underscore, hyphen)"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8,
                    "description": "Password (must include uppercase, lowercase, digit, special char)"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "Email address (optional)"
                  },
                  "public_key": {
                    "type": "string",
                    "description": "Pre-existing public key (optional)"
                  },
                  "private_key": {
                    "type": "string",
                    "description": "Pre-existing private key (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user_id": {
                      "type": "integer"
                    },
                    "api_key": {
                      "type": "string",
                      "description": "API key (only returned once - store it securely!)"
                    },
                    "public_key": {
                      "type": "string",
                      "description": "ASCII-armored public key"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["User Management"],
        "summary": "User login",
        "description": "Authenticate a user with username and password. Note: API key is NOT returned on login.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user_id": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/register/form": {
      "post": {
        "tags": ["User Management"],
        "summary": "Register user (form-based)",
        "description": "Register a new user with form data (multipart/form-data). Supports file uploads for existing PGP keys.",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "public_key_file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Public key file (optional)"
                  },
                  "public_key_text": {
                    "type": "string",
                    "description": "Public key as text (optional)"
                  },
                  "private_key_file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Private key file (optional)"
                  },
                  "private_key_text": {
                    "type": "string",
                    "description": "Private key as text (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user_id": {
                      "type": "integer"
                    },
                    "api_key": {
                      "type": "string",
                      "description": "API key (only returned once)"
                    },
                    "public_key": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Registration failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/get_api_key": {
      "post": {
        "tags": ["User Management"],
        "summary": "Get API key (masked)",
        "description": "Retrieve masked API key with password authentication. Full key cannot be retrieved after registration.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user_id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    },
                    "api_key_masked": {
                      "type": "string",
                      "description": "Masked API key"
                    },
                    "note": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/profile": {
      "get": {
        "tags": ["User Management"],
        "summary": "Get user profile",
        "description": "Retrieve authenticated user's profile information",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user_id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "has_public_key": {
                      "type": "boolean"
                    },
                    "api_key_masked": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["User Management"],
        "summary": "Update user profile",
        "description": "Update user profile information",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user_id": {
                      "type": "integer"
                    },
                    "username": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/keys/download": {
      "get": {
        "tags": ["User Management"],
        "summary": "Download PGP keys",
        "description": "Download user's public or private PGP key",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Key type to download",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["public", "private"],
              "default": "public"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PGP key file",
            "content": {
              "application/pgp-keys": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/keys/upload": {
      "post": {
        "tags": ["User Management"],
        "summary": "Upload PGP keys",
        "description": "Upload new PGP keys (requires password confirmation)",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["password"],
                "properties": {
                  "password": {
                    "type": "string",
                    "description": "Password confirmation"
                  },
                  "public_key_file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Public key file (optional)"
                  },
                  "private_key_file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Private key file (optional)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Keys uploaded successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "updated_keys": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Upload failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Invalid password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}